+++ bsn1/src/lib.rs
@@ -74,13 +74,11 @@ pub enum Error {
     /// Note that this error cannot be compared with others.
     /// `PartialEq::eq` always returns `false` for this error.
     Io(std::io::Error),
-    /// Wrapper of [`std::error::Error`].
-    ///
-    /// This is used when users want to use their own error type.
+    /// Wrapper of [`anyhow::Error`].
     ///
     /// Note that this error cannot be compared with others.
     /// `PartialEq::eq` always returns `false` for this error.
-    Boxed(Box<dyn std::error::Error>),
+    Anyhow(anyhow::Error),
 }
 
 impl fmt::Display for Error {
@@ -96,7 +94,7 @@ impl fmt::Display for Error {
             Self::InvalidUtf8 => f.write_str("Invalid as UTF-8."),
             Self::InvalidKeyValuePair => f.write_str("The key-value pair is invalid."),
             Self::Io(err) => err.fmt(f),
-            Self::Boxed(err) => err.fmt(f),
+            Self::Anyhow(err) => err.fmt(f),
         }
     }
 }
@@ -116,7 +114,7 @@ impl PartialEq for Error {
             Self::InvalidUtf8 => matches!(other, Self::InvalidUtf8),
             Self::InvalidKeyValuePair => matches!(other, Self::InvalidKeyValuePair),
             Self::Io(_) => false,
-            Self::Boxed(_) => false,
+            Self::Anyhow(_) => false,
         }
     }
 }
@@ -127,11 +125,57 @@ impl From<std::io::Error> for Error {
     }
 }
 
-impl<T> From<Box<T>> for Error
-where
-    T: 'static + std::error::Error,
-{
-    fn from(err: Box<T>) -> Self {
-        Self::Boxed(err as Box<dyn std::error::Error>)
+impl From<anyhow::Error> for Error {
+    fn from(err: anyhow::Error) -> Self {
+        Self::Anyhow(err)
+    }
+}
+
+impl Error {
+    /// Consumes `self`, converting it into an [`anyhow::Error`].
+    ///
+    /// If `self` matches `Error::Anyhow`, returns the inner value;
+    /// otherwise, wraps `self` and returns.
+    pub fn into_anyhow(self) -> anyhow::Error {
+        match self {
+            Self::Anyhow(err) => err,
+            _ => anyhow::Error::new(self),
+        }
+    }
+
+    /// Returns a reference to the inner [`anyhow::Error`] if `self` is `Error::Anyhow`;
+    /// otherwise, returns `None`.
+    pub fn as_anyhow(&self) -> Option<&anyhow::Error> {
+        match self {
+            Self::Anyhow(err) => Some(err),
+            _ => None,
+        }
+    }
+
+    /// Consumes `self`, wrapping it with `context`.
+    pub fn context<C>(self, context: C) -> Self
+    where
+        C: fmt::Display + Send + Sync + 'static,
+    {
+        self.into_anyhow().context(context).into()
+    }
+
+    /// If `self` matches `Error::Anyhow`, returns a reference to the first `Self` type error
+    /// in the chain; otherwise, returns `self`.
+    pub fn root_cause(&self) -> &Self {
+        match self {
+            Self::Anyhow(err) => {
+                let mut ret = self;
+                for cause in err.chain() {
+                    if let Some(e) = cause.downcast_ref::<Self>() {
+                        ret = e;
+                    } else {
+                        break;
+                    }
+                }
+                return ret;
+            }
+            _ => self,
+        }
     }
 }