+++ bsn1/CHANGELOG.md
@@ -394,6 +394,7 @@ Nothing is changed.
 
 - Add variant `Anyhow` to enum `Error`
 - Implement `From<anyhow::Error>` for `Error`
+- Create method `Error::into\_anyhow()`
 
 ### Removed